+++ karma.conf.js
@@ -116,6 +116,7 @@ module.exports = function(config) {
     coverageReporter: {
       dir: 'coverage',
       reporters: [
+        {type: 'lcovonly', subdir: '.'},
         {type: 'json', subdir: '.', file: 'coverage.json'},
         {type: 'html'},
       ],